--- src/engine/fileOps/shared.h
@@ -1,53 +1,0 @@
-/**
- * Furnace Tracker - multi-system chiptune tracker
- * Copyright (C) 2021-2024 tildearrow and contributors
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
- */
-
-#pragma once
-
-#include "importExport.h"
-
-// MIDI-related
-struct midibank_t {
-  String name;
-  uint8_t bankMsb,
-          bankLsb;
-};
-
-// Reused patch data structures
-
-// SBI and some other OPL containers
-
-struct sbi_t {
-  uint8_t Mcharacteristics,
-          Ccharacteristics,
-          Mscaling_output,
-          Cscaling_output,
-          Meg_AD,
-          Ceg_AD,
-          Meg_SR,
-          Ceg_SR,
-          Mwave,
-          Cwave,
-          FeedConnect;
-};
-
-bool stringNotBlank(String& str);
-// detune needs extra translation from register to furnace format
-uint8_t fmDtRegisterToFurnace(uint8_t&& dtNative);
-
-void readSbiOpData(sbi_t& sbi, SafeReader& reader);
